refactor: share url fixtures with src/shared
