+++ __tests__/src/lib/__request_fixtures.js
@@ -1,4 +1,4 @@
-import { A_URL } from '../../shared/__url_fixtures.js';
+import { A_URL } from '../../../src/shared/__url_fixtures.js';
 
 export const RANDOM_REQUEST_DETAILS = { url: A_URL };
 export const BLOCKING_RESPONSE = {
